권한, 조회 수, 검색어 처리 및 공지사항 말머리 추가

+++ src/main/java/anonymous/homepage/board/web/BoardController.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
@@ -33,20 +34,36 @@ public class BoardController {
         boardVO.setRecordUnit(boardVO.getRecordUnit() == 0 ? 10 : boardVO.getRecordUnit());
         boardVO.setPageUnit(boardVO.getPageUnit() == 0 ? 10 : boardVO.getPageUnit());
         boardVO.setNowPage(boardVO.getNowPage() == 0 ? 1 : boardVO.getNowPage());
-        boardVO.setRecordCount(boardService.selectBoardCount());
+
+        // 검색어 처리
+        String originalSearchKeyword = boardVO.getSearchKeyword();
+        if(StringUtils.hasText(originalSearchKeyword)) {
+            boardVO.setSearchKeyword(originalSearchKeyword.replaceAll(" ", "|"));
+        }
+
+        // 게시판 조회
+        List<BoardVO> boardList = boardService.selectBoardList(boardVO);
+        boardVO.setRecordCount(boardService.selectBoardCount(boardVO));
+        boardVO.setSearchKeyword(originalSearchKeyword);
 
         // 데이터 전달
         model.addAttribute("boardInfo", boardVO);
-        model.addAttribute("boardList", boardService.selectBoardList(boardVO));
+        model.addAttribute("boardList", boardList);
+        model.addAttribute("searchSeCds", cdService.selectCdList("A05"));
+
         return "board/boardList";
     }
 
     // 공지사항 상세 화면 이동
     @GetMapping("/selectBoard.do")
     public String selectBoard(@ModelAttribute BoardVO boardVO, Model model) {
-        // 조회수 증가
-        boardService.updateBoardViewCount(boardVO);
+        // 게시글 조회
         BoardVO board = boardService.selectBoard(boardVO);
+        
+        // 조회수 증가
+        if(!board.getWriter().equals(boardVO.getUserId())) {
+            boardService.updateBoardViewCount(boardVO);
+        }
 
         // 데이터 전달
         model.addAttribute("boardInfo", boardVO);
@@ -79,34 +96,60 @@ public class BoardController {
     // 공지사항 등록
     @PostMapping("/insertBoard.do")
     public String insertBoard(@ModelAttribute BoardVO boardVO, RedirectAttributes redirect) {
-        boardService.insertBoard(boardVO);
+        String resultMessage;
+        String redirectUrl;
 
-        boardVO.setResultMessage("게시글이 등록되었습니다.");
+        if(boardVO.getIsPermitted()) {
+            boardService.insertBoard(boardVO);
+            resultMessage = "게시글이 등록되었습니다.";
+            redirectUrl = "/board/selectBoard.do";
+
+        } else {
+            resultMessage = "권한이 없습니다.";
+            redirectUrl = "/board/selectBoardList.do";
+        }
+
+        boardVO.setResultMessage(resultMessage);
         redirect.addFlashAttribute("boardVO", boardVO);
-
-        return "redirect:/board/selectBoard.do";
+        return "redirect:" + redirectUrl;
     }
 
     // 공지사항 수정
     @PostMapping("/updateBoard.do")
     public String updateBoard(@ModelAttribute BoardVO boardVO, RedirectAttributes redirect) {
-        boardService.updateBoard(boardVO);
+        String resultMessage;
+        String redirectUrl;
 
-        boardVO.setResultMessage("게시글이 수정되었습니다.");
+        if(boardVO.getIsPermitted()) {
+            boardService.updateBoard(boardVO);
+            resultMessage = "게시글이 수정되었습니다.";
+            redirectUrl = "/board/selectBoard.do";
+
+        } else {
+            resultMessage = "권한이 없습니다.";
+            redirectUrl = "/board/selectBoardList.do";
+        }
+
+        boardVO.setResultMessage(resultMessage);
         redirect.addFlashAttribute("boardVO", boardVO);
-
-        return "redirect:/board/selectBoard.do";
+        return "redirect:" + redirectUrl;
     }
 
     // 공지사항 삭제
     @PostMapping("/deleteBoard.do")
     public String deleteBoard(@ModelAttribute BoardVO boardVO, RedirectAttributes redirect) {
-        boardService.deleteBoard(boardVO);
+        String resultMessage;
+        if(boardVO.getIsPermitted()) {
+            boardService.deleteBoard(boardVO);
+            resultMessage = "게시글이 삭제되었습니다.";
+            boardVO.setNowPage(1);
 
-        boardVO.setResultMessage("게시글이 삭제되었습니다.");
-        boardVO.setNowPage(1);
+        } else {
+            resultMessage = "권한이 없습니다.";
+        }
+
+        boardVO.setResultMessage(resultMessage);
         redirect.addFlashAttribute("boardVO", boardVO);
-
         return "redirect:/board/selectBoardList.do";
     }
